IGNITE-13492: fix parallel
